Notebooks/Streamlit/requirements.txt
streamlit==1.8.0
boto3==1.18.48
fastapi==0.68.0
uvicorn==0.15.0
snowflake-connector-python
